+++ docs/form_qc_checker/CHANGELOG.md
@@ -3,8 +3,10 @@
 All notable changes to this gear are documented in this file.
 
 ## 1.5.6
+* Changes to use the ADCID from the parent project custom info for the pipeline ADCID. Fails if the ADCID is not set.
+* Rebuilt for ssm-parameter-store update
 * Rebuilt for new preprocessing checks in config
-
+  
 ## 1.5.5
 * Updates error metadata to include date and naccid